Fix warnings of the form   warning: long int format, different type arg (arg 5) by casting ino_t arguments to unsigned long for printf formats. In some instances, change %ld to %lu.

+++ fs/nfsd/export.c
@@ -781,7 +781,8 @@ exp_unexport(struct nfsctl_export *nxp)
 			expkey_put(&key->h, &svc_expkey_cache);
 		} else
 			dprintk("nfsd: no export %x/%lx for %s\n",
-				nxp->ex_dev, nxp->ex_ino, nxp->ex_client);
+				nxp->ex_dev, (unsigned long) nxp->ex_ino,
+				nxp->ex_client);
 		auth_domain_put(dom);
 		cache_flush();
 	} else
